backend/pyproject.toml
[project]
name = "Examenia"
version = "0.1.0"
description = "The projec for attempting exam using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clerk-backend-api>=2.2.0",
    "fastapi>=0.115.12",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
